modularizes npmdata/plugindata
