+++ src/test/suite/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable arrow-parens */
 import * as path from 'path';
 import * as Mocha from 'mocha';
 import * as glob from 'glob';
